Make presentation adapter so it can be used for any use case such as loading comments, images, feed, etc...
