--- ZZFeedApp/ZZFeedApp/Composers/FeedLoaderPresentationAdapter.swift
@@ -1,31 +1,0 @@
-//
-//  Copyright © zzmasoud (github.com/zzmasoud).
-//  
-
-import Foundation
-import Combine
-import ZZFeed
-import ZZFeediOS
-
-final class FeedLoaderPresentationAdapter: FeedViewControllerDelegate {
-    private let feedLoader: () -> FeedLoader.Publisher
-    private var cancellable: Cancellable?
-    var presenter: LoadResourcePresenter<[FeedItem], FeedViewAdapter>?
-    
-    init(feedLoader: @escaping () -> FeedLoader.Publisher) {
-        self.feedLoader = feedLoader
-    }
-    
-    func didRequestFeedRefresh() {
-        presenter?.didStartLoading()
-        
-        cancellable = feedLoader().sink { [weak self] completion in
-            if case let .failure(error) = completion {
-                self?.presenter?.didFinishLoading(with: error)
-            }
-        } receiveValue: { [weak self] feed in
-            self?.presenter?.didFinishLoading(with: feed)
-        }
-
-    }
-}
