+++ ZZFeedApp/ZZFeedApp/Composers/LoadResourcePresentationAdapter.swift
@@ -1,0 +1,38 @@
+//
+//  Copyright © zzmasoud (github.com/zzmasoud).
+//  
+
+import Foundation
+import Combine
+import ZZFeed
+import ZZFeediOS
+
+final class LoadResourcePresentationAdapter<Resource, View: ResourceView> {
+    typealias LoadPublisher = AnyPublisher<Resource, Error>
+    
+    private let loader: () -> LoadPublisher
+    private var cancellable: Cancellable?
+    var presenter: LoadResourcePresenter<Resource, View>?
+    
+    init(loader: @escaping () -> LoadPublisher) {
+        self.loader = loader
+    }
+    
+    func loadResource() {
+        presenter?.didStartLoading()
+        
+        cancellable = loader().sink { [weak self] completion in
+            if case let .failure(error) = completion {
+                self?.presenter?.didFinishLoading(with: error)
+            }
+        } receiveValue: { [weak self] resource in
+            self?.presenter?.didFinishLoading(with: resource)
+        }
+    }
+}
+
+extension LoadResourcePresentationAdapter: FeedViewControllerDelegate {
+    func didRequestFeedRefresh() {
+        loadResource()
+    }
+}
